♻️(app-desk) load domain  mailboxes in MailDomainsContent
update pages/mail-domains/[name].tsx to not fetch mailboxes
update features/mail-domains/components/MailDomainsContent.tsx

to fetch mailboxes with useMailDomainMailboxes hook
